USB support added and support for more CPUs
Former-commit-id: c427ddc0812ea4ed0ef62d90ff924eb891b1be2e

+++ hardware/pic32/cores/pic32/pins_arduino_pic32_cerebot32mx4.cxx
@@ -173,7 +173,7 @@ const uint8_t	digital_pin_to_port_PGM[] = {
 };
 
 //************************************************************************
-const uint32_t PROGMEM digital_pin_to_bit_mask_PGM[] = 
+const uint16_t PROGMEM digital_pin_to_bit_mask_PGM[] = 
 {
 	// PIN IN PORT
 	// -------------------------------------------
